labs/pipeline-generator-cdk/requirements.txt
-e .
aws_cdk.aws_codebuild
aws_cdk.aws_codecommit
aws_cdk.aws_codepipeline
aws_cdk.aws_codepipeline_actions
aws_cdk.aws_ec2
aws_cdk.aws_ecs
aws_cdk.aws_ecs_patterns